// packages/40009/src copy/apis/register.ts
export interface DeptType {
  createTime: string;
  employeeCount: number;
  hisId: number;
  hisType: number;
  id: number;
  isSummary: number;
  name: string;
  no: string;
  pathCode: string;
  pid: number;
  sortNo: number;
  status: number;
  tel: string;
  updateTime: string;
  address?: string;
  children: DeptType[];
  img: string;
  summary: string;
}
// ...
export interface RegisterConfimType {
  deptName: string;
  doctorName: string;
  doctorTitle: string;
  hisName: string;
  leftBindNum: number;
  patientList: PatientType[];
  registerType: string;
  registerTypeName: string;
  scheduleDate: string;
  totalFee: number;
  visitBeginTime: string;
  visitEndTime: string;
  visitPeriod: 1 | 2 | 3 | 4 | 5;
  visitWeekName: string;
}
// ...
export interface DoctorDetailType {
  circleImage: string;
  deptId: number;
  deptImage: string;
  deptName: string;
  deptNo: string;
  doctorId: string;
  grade: string;
  hisDoctorId: string;
  hisDoctorName: string;
  hisId: number;
  hisName: string;
  hisType: number;
  id: number;
  image: string;
  introduction: string;
  isDeleted: number;
  level: string;
  name: string;
  qrContent: string;
  qrTicket: string;
  qrUrl: string;
  specialty: string;
  status: string;
  type: string;
  workingLife: number;
  hisDistrict: string;
}