// packages/40009/src copy/apis/microsite.ts
export interface DeptDetailType {
  address?: string;
  createTime?: string;
  hisDistrict?: string;
  hisId?: number;
  hisType?: number;
  hospitalDeptName?: string;
  hospitalDeptNo?: string;
  id?: number;
  img?: string;
  initials?: string;
  isDeleted?: number;
  medicalDepartment?: string;
  name?: string;
  no?: string;
  pathCode?: string;
  pid?: number;
  skill?: string;
  standardDeptName?: null;
  standardDeptNo?: null;
  status?: number;
  summary?: string;
  tel?: string;
  updateTime?: string;
}